added interface to services
