+++ src/main/java/edutech/backend/entity/RefreshToken.java
@@ -1,10 +1,7 @@
 package edutech.backend.entity;
 
-
-
 import jakarta.persistence.*;
 import lombok.Data;
-
 import java.time.Duration;
 import java.time.Instant;
