+++ src/main/java/edutech/backend/service/CourseServiceImpl.java
@@ -1,0 +1,66 @@
+package edutech.backend.service;
+
+import edutech.backend.entity.Course;
+import edutech.backend.exception.CourseNotFoundException;
+import edutech.backend.repository.CourseRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import java.util.List;
+import java.util.Optional;
+
+@Service
+public class CourseServiceImpl implements CourseService{
+
+    @Autowired
+    private CourseRepository courseRepository;
+
+    // Get all courses from the database
+    @Transactional(readOnly = true)
+    @Override
+    public List<Course> getAllCourses() {
+        return courseRepository.findAll();
+    }
+
+    // Get a course by ID
+    @Transactional(readOnly = true)
+    @Override
+    public Course findCourseById(Long id) {
+        // Fetch the course from the repository
+        Optional<Course> courseOptional = courseRepository.findById(id);
+        // If the course is present, return it. Otherwise, throw an exception
+        return courseOptional.orElseThrow(() -> new CourseNotFoundException("Course not found with id: " + id));
+    }
+
+    // Create a new course in the database
+    @Transactional
+    @Override
+    public Course createCourse(Course course) {
+        return courseRepository.save(course);
+    }
+
+    // Update an existing course by ID
+    @Transactional
+    @Override
+    public Course updateCourse(Long id, Course updatedCourse) {
+        // Fetch the existing course from the database
+        Course existingCourse = courseRepository.findById(id)
+                .orElseThrow(() -> new CourseNotFoundException("Course not found with id: " + id));
+
+        // Update course fields with the new data
+        existingCourse.setName(updatedCourse.getName());
+        existingCourse.setDescription(updatedCourse.getDescription());
+
+        // Save the updated course back into the database
+        return courseRepository.save(existingCourse);
+    }
+
+    // Delete a course by its ID
+    @Transactional
+    @Override
+    public String deleteCourseById(Long id) {
+        // Try to delete the course by its ID
+        courseRepository.deleteById(id);
+        return "Course deleted successfully.";
+    }
+}
